refactor: Move sketch solver interop code to sketch_solve module
